refactor: make database herb provenance canonical
